Disabled SELinux in default kickstart template. Added check in rpower softoff to check if node is pingable, if not proceed to shutdown instead of waiting. Rearranged buttons on action bar in self-service page. Added shutdown button in nodes page to shutdown node gracefully.
git-svn-id: https://svn.code.sf.net/p/xcat/code/xcat-core/branches/2.8@16168 8638fb3e-16cb-4fca-ae20-7b5d299a9bcd

+++ xCAT-server/lib/xcat/plugins/zvm.pm
@@ -2305,8 +2305,10 @@ sub powerVM {
     
     # Power off virtual server (gracefully)
     elsif ( $args->[0] eq 'softoff' ) {
-        $out = `ssh -o ConnectTimeout=10 $::SUDOER\@$node "shutdown -h now"`;
-        sleep(15);    # Wait 15 seconds before logging user off
+    	if (`pping $node` !~ m/noping/i) {
+            $out = `ssh -o ConnectTimeout=10 $::SUDOER\@$node "shutdown -h now"`;
+            sleep(15);    # Wait 15 seconds before logging user off
+    	}
         
         $out = `ssh $::SUDOER\@$hcp "$::SUDO $::DIR/smcli Image_Deactivate -T $userId"`;
         xCAT::zvmUtils->printSyslog("smcli Image_Deactivate -T $userId");
@@ -3960,6 +3962,9 @@ sub clone {
         push( @tgtDisks, $addr );
         $type       = $words[2];
         $mode       = $words[6];
+        if (!$mode) {
+        	$mode = "MR";
+        }
         
         # Add 0 in front if address length is less than 4
         while (length($addr) < 4) {
